+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## \[Unreleased\]
+### New features
+### Enhancements
+- Enhance import performance for built-in plugins
+  (<https://github.com/openvinotoolkit/datumaro/pull/1031>)
+### Bug fixes
 
 ## 26/05/2023 - Release 1.3.1
 ### Bug fixes